Treesame-Commit-Id: 04141e5a85092f104d9e9e73037e3bb6706c646c

+++ db/migrate/20160322204834_drop_zip_file_imports.rb
@@ -3,7 +3,7 @@ class DropZipFileImports < ActiveRecord::Migration
   disable_ddl_transaction!
 
   def up
-    drop_table :zip_file_imports
+    drop_table :zip_file_imports if connection.table_exists?(:zip_file_imports)
 
     if Attachment.where(:context_type => "ZipFileImport").exists?
       Attachment.where(:context_type => "ZipFileImport").
